Extracted state to a separate file.
