+++ packages/ckeditor5-find-and-replace/src/findandreplaceediting.js
@@ -14,86 +14,14 @@ import ReplaceCommand from './replacecommand';
 import ReplaceAllCommand from './replaceallcommand';
 import FindNextCommand from './findnextcommand';
 import FindPreviousCommand from './findpreviouscommand';
+import FindAndReplaceState from './findandreplacestate';
 
-import { ObservableMixin, mix, Collection } from 'ckeditor5/src/utils';
 // eslint-disable-next-line ckeditor5-rules/ckeditor-imports
 import { scrollViewportToShowTarget } from '@ckeditor/ckeditor5-utils/src/dom/scroll';
 
 import { debounce } from 'lodash-es';
 
 import '../theme/findandreplace.css';
-
-/**
- * The object storing find & replace plugin state in a given editor instance.
- *
- */
-class FindAndReplaceState {
-	constructor( model ) {
-		this.set( 'results', new Collection() );
-
-		this.set( 'highlightedResult', null );
-
-		this.set( 'searchText', '' );
-
-		this.set( 'replaceText', '' );
-
-		this.set( 'matchCase', false );
-		this.set( 'matchWholeWords', false );
-
-		this.results.on( 'change', ( eventInfo, { removed, index } ) => {
-			removed = Array.from( removed );
-
-			if ( removed.length ) {
-				let highlightedResultRemoved = false;
-
-				model.change( writer => {
-					for ( const removedResult of removed ) {
-						if ( this.highlightedResult === removedResult ) {
-							highlightedResultRemoved = true;
-						}
-
-						if ( model.markers.has( removedResult.marker.name ) ) {
-							writer.removeMarker( removedResult.marker );
-						}
-					}
-				} );
-
-				if ( highlightedResultRemoved ) {
-					const nextHighlightedIndex = index >= this.results.length ? 0 : index;
-					this.highlightedResult = this.results.get( nextHighlightedIndex );
-				}
-			}
-		} );
-	}
-
-	clear( model ) {
-		// @todo: actually this handling might be moved to editing part.
-		// This could be a results#change listener that would ensure that related markers are ALWAYS removed without
-		// having to call state.clear() explicitly.
-
-		this.searchText = '';
-		// this.replaceText = '';
-
-		model.change( writer => {
-			if ( this.highlightedResult ) {
-				const oldMatchId = this.highlightedResult.marker.name.split( ':' )[ 1 ];
-				const oldMarker = model.markers.get( `findResultHighlighted:${ oldMatchId }` );
-
-				if ( oldMarker ) {
-					writer.removeMarker( oldMarker );
-				}
-			}
-
-			[ ...this.results ].forEach( ( { marker } ) => {
-				writer.removeMarker( marker );
-			} );
-		} );
-
-		this.results.clear();
-	}
-}
-
-mix( FindAndReplaceState, ObservableMixin );
 
 const HIGHLIGHT_CLASS = 'ck-find-result_selected';
